refactor(api/v0): centralize DTOs into internal/api/v0/dto package
Each DTO type now lives in its own file with its dedicated mapper.
Eliminates duplication of RunRecordDTO/ReturnDTO between ws and http
handlers, unifying ReturnDTO to include Variables (omitempty).
Old dtos.go files removed; tests migrated to dto/ with TestMain
relocated to the surviving handler test files.
