+++ internal/api/v0/dto/execute_method_request.go
@@ -1,0 +1,5 @@
+package dto
+
+type ExecuteMethodRequestDTO struct {
+	Variables map[string]string `json:"variables"`
+}
